+++ lib/src/lexeme.dart
@@ -1,5 +1,6 @@
 import 'package:token_parser/src/error.dart';
 import 'package:token_parser/src/grammar.dart';
+import 'package:token_parser/src/lexemes/until.dart';
 import 'package:token_parser/src/token.dart';
 
 /* -= Lexeme Imports =- */
@@ -21,6 +22,7 @@ import 'package:token_parser/src/lexemes/start.dart';
 import 'package:token_parser/src/lexemes/end.dart';
 
 import 'package:token_parser/src/lexemes/character.dart';
+import 'package:token_parser/src/lexemes/repeat.dart';
 
 import 'package:token_parser/src/lexemes/reference.dart';
 import 'package:token_parser/src/lexemes/self.dart';
@@ -191,6 +193,8 @@ abstract class Lexeme extends Pattern {
 
   factory Lexeme.character(Pattern pattern, { String? name }) = CharacterLexeme;
   factory Lexeme.any() = CharacterLexeme.any;
+  factory Lexeme.until(Pattern pattern, Pattern until, { String? name }) = UntilLexeme;
+  factory Lexeme.repeat(Pattern pattern, int min, { int? max, String? name }) = RepeatLexeme;
 
   factory Lexeme.reference(String name, { Grammar? grammar }) = ReferenceLexeme;
   factory Lexeme.ref(String name, { Grammar? grammar }) = ReferenceLexeme;