+++ lib/src/lexemes/and.dart
@@ -24,7 +24,7 @@ class AndLexeme extends ParentLexeme {
   String get regexString => children.isNotEmpty
     ? children.length == 1
       ? children.first.toString()
-      : '(?:${children.join('')})'
+      : '(?:${children.map((token) => token.regexString).join('')})'
     : '';
 }